+++ src/Resources/contao/config/config.php
@@ -1,6 +1,11 @@
 <?php
 
-$GLOBALS['TL_HOOKS']['formhybridOnCreateInstance'][]            = [
-    \HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridOnCreateInstanceListener::class, '__invoke'];
-$GLOBALS['TL_HOOKS'][\HeimrichHannot\FormHybrid\Event\FormhybridBeforeCreateWidgetEvent::NAME][] = [
-    \HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridBeforeCreateWidgetEventListener::class, '__invoke'];
+use HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridBeforeCreateWidgetEventListener;
+use HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridModifyAsyncFormSubmitResponseEventListener;
+use HeimrichHannot\ContaoFormhybridCompatibilityBundle\EventListener\FormhybridOnCreateInstanceListener;
+use HeimrichHannot\FormHybrid\Event\FormhybridBeforeCreateWidgetEvent;
+use HeimrichHannot\FormHybrid\Event\FormhybridModifyAsyncFormSubmitResponseEvent;
+
+$GLOBALS['TL_HOOKS']['formhybridOnCreateInstance'][]                       = [FormhybridOnCreateInstanceListener::class, '__invoke'];
+$GLOBALS['TL_HOOKS'][FormhybridBeforeCreateWidgetEvent::NAME][]            = [FormhybridBeforeCreateWidgetEventListener::class, '__invoke'];
+$GLOBALS['TL_HOOKS'][FormhybridModifyAsyncFormSubmitResponseEvent::NAME][] = [FormhybridModifyAsyncFormSubmitResponseEventListener::class, '__invoke'];